SEQUOIADBMAINSTREAM-3650 代码编译告警清理
git-svn-id: http://192.168.20.11/sequoiadb/trunk@35991 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/coord/coordCommandNode.cpp
@@ -1904,7 +1904,7 @@ namespace engine
                ++validCount ;
             }
 
-            if ( pArgs->_boQuery.nFields() > validCount )
+            if ( (UINT32)pArgs->_boQuery.nFields() > validCount )
             {
                rc = SDB_INVALIDARG ;
                PD_LOG( PDERROR, "Unknown parameters in command's args[%s]",
@@ -2318,7 +2318,7 @@ namespace engine
             }
          }
 
-         if ( pArgs->_boQuery.nFields() > validCount )
+         if ( (UINT32)pArgs->_boQuery.nFields() > validCount )
          {
             rc = SDB_INVALIDARG ;
             PD_LOG( PDERROR, "Unknown parameters in command's args[%s]",
